🐛🤖 always include the not-applicable pattern
The pattern was only emitted when the rendered view happened to hold a
not-applicable country, but the legend bin is derived from the whole
table, so a faceted or zoomed view without one drew an unpainted legend
swatch. It now sits in the chart-wide defs next to the no-data pattern;
only the viewport-scaled variant stays local to the 2D map.
