+++ packages/@ourworldindata/grapher/src/mapCharts/MapComponents.tsx
@@ -174,32 +174,6 @@ export function NoDataPattern({
     )
 }
 
-export function InapplicablePattern({
-    patternId,
-    scale = 1,
-}: {
-    patternId: string
-    scale?: number
-}): React.ReactElement {
-    const roundedScale = R.round(scale, 3)
-    return (
-        <pattern
-            id={patternId}
-            patternUnits="userSpaceOnUse"
-            width="4"
-            height="4"
-            patternTransform={`rotate(-45 2 2) scale(${roundedScale})`}
-        >
-            <rect width="4" height="4" fill="#ccc" />
-            <path
-                d="M -1,2 l 6,0"
-                stroke="#fff"
-                strokeWidth={PATTERN_STROKE_WIDTH}
-            />
-        </pattern>
-    )
-}
-
 export function ProjectedDataPattern({
     color,
     scale = 1,
